Store sidebarHidden state in redux store
This saves us a couple of callbacks and state handling in multiple
components. The components that need to be aware of the sidebar
state simply use the "mapStateToProps" function to retrieve the
value from the redux store ("navigation.sidebarHidden").

Furthermore, this changes the inheritance relationship between
*SettingsBar to Sidebar to a composition relationship instead.
This saves us the need to store the header element in a state
variable. In React, the state should be kept minimal.

This commit closes #4.